@@ -1,83 +1,82 @@
 {
-    "name": "item-refactor",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve --proxy-config ./proxy.json",
-        "server": "ts-node -P ./server/server.tsconfig.json ./server/server.ts",
-        "watch": "nodemon -w . --ext \".ts\" --exec \"npm run start\"",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular-devkit/schematics": "^8.3.24",
-        "@angular/animations": "~8.2.14",
-        "@angular/cdk": "^8.2.3",
-        "@angular/common": "~8.2.14",
-        "@angular/compiler": "~8.2.14",
-        "@angular/core": "~8.2.14",
-        "@angular/forms": "~8.2.14",
-        "@angular/material": "^8.2.3",
-        "@angular/material-moment-adapter": "^8.2.3",
-        "@angular/platform-browser": "~8.2.14",
-        "@angular/platform-browser-dynamic": "~8.2.14",
-        "@angular/router": "~8.2.14",
-        "@ng-bootstrap/ng-bootstrap": "^5.2.1",
-        "@ng-bootstrap/schematics": "^2.0.0-alpha.1",
-        "@ngrx/data": "^8.6.0",
-        "@ngrx/effects": "^8.6.0",
-        "@ngrx/entity": "^8.6.0",
-        "@ngrx/router-store": "^8.6.0",
-        "@ngrx/store": "^8.6.0",
-        "@ngrx/store-devtools": "^8.6.0",
-        "@ngx-translate/core": "^12.0.0",
-        "@ngx-translate/http-loader": "^4.0.0",
-        "body-parser": "^1.19.0",
-        "bootstrap": "^4.4.1",
-        "core-js": "^3.6.4",
-        "express": "^4.17.1",
-        "hammerjs": "^2.0.8",
-        "jasmine-marbles": "^0.6.0",
-        "moment": "^2.24.0",
-        "ngx-cookie-service": "^2.3.0",
-        "node-gyp": "^6.1.0",
-        "rxjs": "~6.4.0",
-        "tslib": "^1.10.0",
-        "zone.js": "~0.9.1"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.803.23",
-        "@angular/cli": "~8.3.23",
-        "@angular/compiler-cli": "~8.2.14",
-        "@angular/language-service": "~8.2.14",
-        "@ngrx/schematics": "^8.6.0",
-        "@types/bluebird": "^3.0.35",
-        "@types/express": "^4.17.2",
-        "@types/jasmine": "~3.3.8",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~8.9.4",
-        "@types/sequelize": "4.0.38",
-        "@typescript-eslint/eslint-plugin": "^2.18.0",
-        "@typescript-eslint/parser": "^2.18.0",
-        "codelyzer": "^5.0.0",
-        "eslint": "^6.8.0",
-        "eslint-config-prettier": "^6.9.0",
-        "eslint-plugin-html": "^6.0.0",
-        "eslint-plugin-prettier": "^3.1.2",
-        "jasmine-core": "~3.4.0",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.1.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~2.0.1",
-        "karma-jasmine-html-reporter": "^1.4.0",
-        "lodash": "^4.17.15",
-        "prettier": "^1.19.1",
-        "protractor": "~5.4.0",
-        "ts-node": "^8.6.2",
-        "typescript": "^3.5.3"
-    }
+  "name": "item-refactor",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --proxy-config ./proxy.json",
+    "server": "ts-node -P ./server/server.tsconfig.json ./server/server.ts",
+    "watch": "nodemon -w . --ext \".ts\" --exec \"npm run start\"",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular-devkit/schematics": "^9.0.1",
+    "@angular/animations": "~9.0.0",
+    "@angular/cdk": "^9.0.0",
+    "@angular/common": "~9.0.0",
+    "@angular/compiler": "~9.0.0",
+    "@angular/core": "~9.0.0",
+    "@angular/forms": "~9.0.0",
+    "@angular/material": "^9.0.0",
+    "@angular/material-moment-adapter": "^9.0.0",
+    "@angular/platform-browser": "~9.0.0",
+    "@angular/platform-browser-dynamic": "~9.0.0",
+    "@angular/router": "~9.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.2.1",
+    "@ng-bootstrap/schematics": "^2.0.0-alpha.1",
+    "@ngrx/data": "^8.6.0",
+    "@ngrx/effects": "^8.6.0",
+    "@ngrx/entity": "^8.6.0",
+    "@ngrx/router-store": "^8.6.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngrx/store-devtools": "^8.6.0",
+    "@ngx-translate/core": "^12.0.0",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "body-parser": "^1.19.0",
+    "bootstrap": "^4.4.1",
+    "core-js": "^3.6.4",
+    "express": "^4.17.1",
+    "jasmine-marbles": "^0.6.0",
+    "moment": "^2.24.0",
+    "ngx-cookie-service": "^2.3.0",
+    "node-gyp": "^6.1.0",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular/cli": "~9.0.1",
+    "@angular/compiler-cli": "~9.0.0",
+    "@angular/language-service": "~9.0.0",
+    "@ngrx/schematics": "^8.6.0",
+    "@types/bluebird": "^3.0.35",
+    "@types/express": "^4.17.2",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "@types/sequelize": "4.0.38",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
+    "codelyzer": "^5.1.2",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-plugin-html": "^6.0.0",
+    "eslint-plugin-prettier": "^3.1.2",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "lodash": "^4.17.15",
+    "prettier": "^1.19.1",
+    "protractor": "~5.4.0",
+    "ts-node": "^8.6.2",
+    "typescript": "^3.7.5"
+  }
 }
