@@ -16,7 +16,7 @@ import { NavComponent } from './nav/nav.component';
     ]
 })
 export class CoreModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<CoreModule> {
         return {
             ngModule: CoreModule,
             providers: [
